Algoritmo hecho sin correccion
Por alguna razon usa al reves a veces el orden de los numeros en el momento de calcular combinacion, pero los calculos se demuestran correctos igualmente.

+++ TPOCalcularCombinaciones/src/main/java/Modelo/Combinacion.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 
 public class Combinacion {
     private ArrayList<Operadores> operadores;
-    private ArrayList<Integer> numeros;
+    private ArrayList<Double> numeros;
 
-    public Combinacion(ArrayList<Operadores> operadores, ArrayList<Integer> numeros) {
+    public Combinacion(ArrayList<Operadores> operadores, ArrayList<Double> numeros) {
         this.operadores = operadores;
         this.numeros = numeros;
     }
@@ -15,7 +15,7 @@ public class Combinacion {
         return operadores;
     }
 
-    public ArrayList<Integer> getNumeros() {
+    public ArrayList<Double> getNumeros() {
         return numeros;
     }
 }
